% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.normal.R
\name{check.normal}
\alias{check.normal}
\title{Title}
\usage{
check.normal(x, breaks = 10, ...)
}
\arguments{
\item{x}{vactor of numeric}

\item{breaks}{number of bin}

\item{...}{args passed to nothing}
}
\description{
Title
}
