% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.na.R
\name{find.na}
\alias{find.na}
\title{find NAs in a dataframe & otionaly graph them}
\usage{
find.na(file, plot = F)
}
\arguments{
\item{file}{dataframe in which to check for na}

\item{plot}{logical value. should a graph be reported}
}
\value{
either the posisiton of NAs of a grpah
}
\description{
find NAs in a dataframe & otionaly graph them
}
\examples{
find.na(data,plot=T)
}
