% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.angle.R
\name{g.angle}
\alias{g.angle}
\title{caculate angle between 2 vectors}
\usage{
g.angle(v1, v2, unit = "degree")
}
\arguments{
\item{v1}{vector 1}

\item{v2}{vector 1}

\item{unit}{character,. either "rad" or "degree"}
}
\value{
angle between 2 vector as a single numeric
}
\description{
caculate angle between 2 vectors
}
\examples{
g.angle(c(1,3,4),c(0,0,-4),"degree")

}

