% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.colorPalette.R
\name{g.colorPalette}
\alias{g.colorPalette}
\title{get a color palette}
\usage{
g.colorPalette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{the name of the pallette or either of any of c('?', 'list','palettes','Palettes','Palette','palette')}

\item{n}{filter by n}

\item{type}{filter by type}
}
\description{
get a color palette
}

