% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.makeGif.R
\name{g.makeGif}
\alias{g.makeGif}
\title{makes an animated gif}
\usage{
g.makeGif(it, fun, file, width = 500, heigth = 500, delay = 10,
  loop = 0)
}
\arguments{
\item{it}{numeric vector of values to pass to fun to generate series of frame}

\item{fun}{function taking for sole input it and returning a figure}

\item{file}{file in wich to save the output gif}

\item{width}{size of the gif in px}

\item{heigth}{size of the gif in px}

\item{delay}{delay value to pass on to convert, in millisecand between frame change}

\item{loop}{value to pass on to convert, should the gif loop. if so keep at 0}
}
\value{
a gif image
}
\description{
makes an animated gif
}
