% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.pairs.R
\name{g.pairs}
\alias{g.pairs}
\title{make pairs graph to observe correlations and all}
\usage{
g.pairs(data, bigger.text = 1, ...)
}
\arguments{
\item{data}{a data frame containing the variables to plot}

\item{bigger.text}{a scaling factor for the text}

\item{...}{other parameters to pass oin to plot}
}
\value{
a graphic of pared correlation and histogram
}
\description{
make pairs graph to observe correlations and all
}
\examples{
data(CO2)
g.pairs(CO2)
}
