% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.source.partial.R
\name{g.source.partial}
\alias{g.source.partial}
\title{Title}
\usage{
g.source.partial(file, start = 0, end = NULL, ...)
}
\arguments{
\item{file}{file to source from}

\item{start}{first line}

\item{end}{last line}

\item{...}{}
}
\description{
Title
}
