% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{draw multile ggplot graphs}
\usage{
multiplot(..., plotlist = NULL, ncol)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{of a list of ggplot}

\item{ncol}{numbre of column}
}
\description{
draw multile ggplot graphs
}

