% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{exctract a column from dplyr as a vector}
\usage{
pull(x, y)
}
\arguments{
\item{x}{tibble}

\item{y}{name of the collumn to extract}
}
\value{
a vector
}
\description{
exctract a column from dplyr as a vector
}
