% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.delim.R
\name{write.delim}
\alias{write.delim}
\title{write tab delimited files}
\usage{
write.delim(d, file)
}
\arguments{
\item{d}{data frame to write}

\item{file}{where to save it}
}
\description{
write tab delimited files
}

